refact task controller, add auth middleware

+++ src/controller/task.js
@@ -1,25 +1,32 @@
 const db = require('../models');
 const Task = db.Task;
 
+const Op = require('sequelize');
+const verifyCollection = require('../utils/verifyCollectionById');
+
 const create = async(req,res)=>{
     try {
         const { collectionId } = req.params;
-        console.log(collectionId);
         const { nom,description } = req.body;
         const existingTask = await Task.findOne({where:{nom}});
 
         if (existingTask) {
             res.status(409).json({
                 message:"Tache déjè existé",
-                data:existingTask
             })
         } else {
-            const task = await Task.create({nom,description,collectionId});
-            await task.save();
-            res.status(201).json({
-                message:"Tache bien ajouté",
-                data:task
-            })
+            if (await verifyCollection(collectionId)) {
+                const task = await Task.create({nom,description,collectionId});
+                await task.save();
+                res.status(201).json({
+                    message:"Tache bien ajouté",
+                    data:task
+                })
+            } else {
+                res.status(404).json({
+                    message:"Collection invalid",
+                })
+            }
         }
     } catch (error) {
         throw new Error(error);
@@ -28,16 +35,23 @@ const create = async(req,res)=>{
 
 const getAll = async(req,res)=>{
     try {
-        const tasks = await Task.findAll({});
-        if (tasks.length>0) {
-            res.status(200).json({
-                message:"Tous les taches",
-                data:tasks
-            })
+        const { collectionId } = req.params;
+        if (await verifyCollection(collectionId)) {
+            const tasks = await Task.findAll({where:{collectionId}});
+            if (tasks.length>0) {
+                res.status(200).json({
+                    message:"Tous les taches",
+                    data:tasks
+                })
+            } else {
+                res.status(404).json({
+                    message:"Tache vide",
+                    data:[]
+                })
+            }
         } else {
-            res.status(409).json({
-                message:"Tache vide",
-                data:[]
+            res.status(404).json({
+                message:"Collection invalid",
             })
         }
     } catch (error) {
@@ -47,16 +61,27 @@ const getAll = async(req,res)=>{
 
 const getById = async(req,res)=>{
     try {
+        const { collectionId } = req.params;
         const { id } = req.params;
-        const task = await Task.findOne({where:{id}});
-        if (task) {
-            res.status(200).json({
-                message:"Tache",
-                data:task
-            })
+        if (await verifyCollection(collectionId)) {
+            const task = await Task.findOne({
+                where: {
+                    [Op.and]: [{id},{collectionId}]
+                }
+            });
+            if (task) {
+                res.status(200).json({
+                    message:"Tache",
+                    data:task
+                })
+            } else {
+                res.status(404).json({
+                    message:"Identifiant 'id' inconnu"
+                })
+            }
         } else {
             res.status(404).json({
-                message:"Identifiant 'id' inconnu"
+                message:"Collection invalid",
             })
         }
     } catch (error) {
@@ -66,27 +91,38 @@ const getById = async(req,res)=>{
 
 const updateTask = async(req,res)=>{
     try {
+        const { collectionId } = req.params;
         const { id } = req.params;
         const updateValue = req.body;
-        const task = await Task.findOne({where:{id}});
+        if (await verifyCollection(collectionId)) {
+            const task = await Task.findOne({
+                where: {
+                    [Op.and]: [{id},{collectionId}]
+                }
+            });
 
-        const newValue = {};
-        Object.keys(updateValue).forEach(uv =>{
-            if (updateValue[uv]!="") {
-                newValue[`${uv}`] = updateValue[uv];
-            }
-        })
-
-        if (task) {
-            task.set(newValue)
-            await task.save();
-            res.status(200).json({
-                message:"Tache modifié",
-                data:task
+            const newValue = {};
+            Object.keys(updateValue).forEach(uv =>{
+                if (updateValue[uv]!="") {
+                    newValue[`${uv}`] = updateValue[uv];
+                }
             })
+
+            if (task) {
+                task.set(newValue)
+                await task.save();
+                res.status(200).json({
+                    message:"Tache modifié",
+                    data:task
+                })
+            } else {
+                res.status(404).json({
+                    message:"Identifiant 'id' inconnu"
+                })
+            }
         } else {
             res.status(404).json({
-                message:"Identifiant 'id' inconnu"
+                message:"Collection invalid",
             })
         }
     } catch (error) {
@@ -96,17 +132,28 @@ const updateTask = async(req,res)=>{
 
 const deleteTask = async(req,res)=>{
     try {
+        const { collectionId } = req.params;
         const { id } = req.params;
-        const task = await Task.findOne({where:{id}});
+        if (await verifyCollection(collectionId)) {
+            const task = await Task.findOne({
+                where: {
+                    [Op.and]: [{id},{collectionId}]
+                }
+            });
 
-        if (task) {
-            task.destroy({});
-            res.status(200).json({
-                message:"Tache supprimé"
-            })
+            if (task) {
+                task.destroy({});
+                res.status(200).json({
+                    message:"Tache supprimé"
+                })
+            } else {
+                res.status(404).json({
+                    message:"Identifiant 'id' inconnu"
+                })
+            }
         } else {
             res.status(404).json({
-                message:"Identifiant 'id' inconnu"
+                message:"Collection invalid",
             })
         }
     } catch (error) {
